86cym3xbe: Move types to models and DTO's
