+++ app/tests/azure/testBlob.ts
@@ -1,0 +1,2 @@
+export const testBlob: Blob = new Blob(['test'], { type: 'text/plain' });
+export const testFile: File = new File(['meow meow meow'], 'test.txt', { type: 'text/plain' })
